Consolidate token page logic and data
