+++ docs/pages/foundations/tokens/typography.tsx
@@ -12,24 +12,13 @@ import { Text } from '@ag.ds-next/react/text';
 import { tokens, FontSize, Font } from '@ag.ds-next/react/core';
 import { Stack } from '@ag.ds-next/react/stack';
 import { InpageNav } from '@ag.ds-next/react/inpage-nav';
-import { DocumentTitle } from '../../../components/DocumentTitle';
 import { TokenLayout } from '../../../components/TokenLayout';
 import { LineHeightChart } from '../../../components/TokenCharts';
-import { TOKEN_PAGES, getTokensBreadcrumbs } from '../../../content/tokens';
 
 export default function TokensTypographyPage() {
 	return (
 		<>
-			<DocumentTitle
-				title={TOKEN_PAGES.typography.pageTitle}
-				description={TOKEN_PAGES.typography.description}
-			/>
-			<TokenLayout
-				title={TOKEN_PAGES.typography.pageTitle}
-				description={TOKEN_PAGES.typography.description}
-				breadcrumbs={getTokensBreadcrumbs(TOKEN_PAGES.typography)}
-				editPath="/docs/pages/foundations/tokens/typography.tsx"
-			>
+			<TokenLayout slug="typography">
 				<InpageNav
 					title="On this page"
 					links={[
